requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
djangorestframework==3.15.2
drf-yasg==1.21.8
exceptiongroup==1.2.2
Faker==33.3.0
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
model-bakery==1.20.1
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.8.2
